refactor(rust): split integration test suite
Extract shared test harness helpers and separate model, cache, and pipeline coverage into dedicated test files.
